e2e tests for SCTaskComplete view
Add fields to strategic action model

Fix the order of SA updates in tasklist view

Search SA updates since last_deadline within tasklist view

+++ run_functional_tests.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+running_locally=${RUNNING_LOCALLY:-true}
+
 # Bring up the db and mock-sso docker containers
 docker-compose -f docker-compose.yaml -f docker-compose.override.yaml up -d
 
@@ -11,11 +13,13 @@ python defend_data_capture/manage.py testserver \
     --settings=test_settings \
     cypress/fixtures/govDepartment.json cypress/fixtures/user.json \
     cypress/fixtures/supplyChains.json cypress/fixtures/strategicActions.json \
+    cypress/fixtures/strategicActionUpdates.json \
     & echo $! > backend.pid \
-    & (sleep 5 && npx cypress run --headless --browser chrome --config-file false)
+    & (sleep 5 && npx cypress run --headless --browser chrome)
 
-# Kill the application using the pid previously saved
-kill $(cat backend.pid)
-
-# Drop the test database
-docker-compose exec db psql -h localhost -U postgres -c "DROP DATABASE IF EXISTS test_defend"
+if [ "$running_locally" == true ]; then
+    # Kill the application using the pid previously saved
+    kill $(cat backend.pid)
+    # Drop the test database
+    docker-compose exec db psql -h localhost -U postgres -c "DROP DATABASE IF EXISTS test_defend"
+fi
